adding pino logger
