+++ package.json
@@ -24,7 +24,8 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "joi": "^17.2.1",
-    "mongoose": "^5.10.2"
+    "mongoose": "^5.10.2",
+    "pino": "^6.5.1"
   },
   "devDependencies": {
     "dotenv": "^8.2.0"
