+++ src/lib/index.js
@@ -1,19 +1,5 @@
 "use strict";
+const { connectDb } = require("./db");
+const logger = require("./logger");
 
-const mongoose = require("mongoose");
-const config = require("../config");
-// import logger from '../logger'
-
-const connectDb = () => {
-  return mongoose
-    .connect(config.get("DB_URL"), {
-      useUnifiedTopology: true,
-      useNewUrlParser: true,
-      useCreateIndex: true,
-    })
-    .then(() => {
-      console.log("db connected successfully");
-    });
-};
-
-module.exports = { connectDb };
+module.exports = { connectDb, logger };
