+++ src/server.js
@@ -4,7 +4,7 @@ const bodyParser = require("body-parser");
 const config = require("./config");
 const app = express();
 const controller = require("./controller");
-const { connectDb } = require("./lib");
+const { connectDb, logger } = require("./lib");
 app
   .use(bodyParser.urlencoded({ extended: false }))
   .use(bodyParser.json())
@@ -12,6 +12,6 @@ app
   .use((req, res) => res.status(404).send("Route not found"));
 connectDb().then(() =>
   app.listen(config.get("PORT"), () => {
-    console.log(`server started on port ${config.get("PORT")}`);
+    logger().info(`server started on port ${config.get("PORT")}`);
   })
 );
